2026.03.19
postArchive 클래스 삭제 - 자주 사용하지 않거나 장기 보존이 필요한 데이터를 운영 시스템에서 별도 저장소로 옮겨 저장하는 기술인데 이 DB는 보통은 Archive log를 이용해서 복구용으로 쓴다고 해서 좀더 공부가 필요할꺼 같아서 현재 DB에서 삭제 요청 후 30일이 지난 데이터를 스케줄을 통해 삭제 하도록 변경함.
PostDeleteBatch 클래스 - Hard Delete 을 위한 batch 클래스
PostResponseDto - viewCount, likeCount의 변수 타입 수정
PostSearchResponseDto - id의 변수 타입 수정
Post 클래스 - 좋아요 수 증가, 감소, 조회수 증가, 감소 비즈니스 메소드 삭제 - 데이터 중복 증가
PostEvent 클래스 - @Bulider 추가, postIds(게시글 목록) 추가
PostEventConsumer - deletePostBulk 메소드 생성 (30일이 지난 후 데이터를 한번에 처리하기 위한 메소드, HardDelete)
PostEventProducer - Topic 상수화, kafkaTemplate 타입 변경, sendPostDeleteBatchEvent 메소드 추가, 이전 코드 수정
PostLikeProducer,ViewCountProducer - Topic 상수화
PostRepository - findDeletePostIdBefore 메소드 추가 및 쿼리문 생성
PostCommandService - Topic 상수화, Redis 캐시 삭제 순서 변경 (구조가 DB 변경 -> 캐시 삭제 -> kafka 이벤트 생성 순으로 되어야함, 즉시 일관성을 위해서 캐시 삭제가 DB 변경 이후에 바로 이루어져야함)
PostLikeService - Topic 상수화, 회원 확인 메소드 삭제 -> 로그인 사용자 기반으로 요청을 컨트롤러에서 하기 때문에 필요가 없다.

+++ src/main/java/com/example/NoticeBoard/domain/post/dto/PostResponseDto.java
@@ -20,9 +20,9 @@ public class PostResponseDto {
 
     private String content; // 게시글 내용
 
-    private Integer viewCount; // 조회수
+    private Long viewCount; // 조회수
 
-    private Integer likeCount; // 게시글 좋아요 수
+    private Long likeCount; // 게시글 좋아요 수
 
     private PostStatus postStatus; // 게시글 종류(일반 게시글, 비밀 게시글, 삭제된 게시글, 블라인드된 게시글 등)
